+++ persistence/src/commonMain/kotlin/com/wire/kalium/persistence/dao/message/MessageDAO.kt
@@ -143,7 +143,7 @@ interface MessageDAO {
     suspend fun getSearchedConversationMessagePosition(
         conversationId: QualifiedIDEntity,
         messageId: String
-    ): Long
+    ): Int
 
     val platformExtensions: MessageExtensions
 }
